# Aplicação/backend/static_docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PM AI MVP API",
    "description": "\n    ## PM AI MVP - Sistema de Gestão de Projetos com IA\n    \n    API completa para gestão de projetos com funcionalidades avançadas de IA.\n    \n    ### Funcionalidades Principais:\n    \n    * **Autenticação**: Sistema de autenticação OAuth com Google\n    * **Projetos**: Gestão completa de projetos e portfólios\n    * **Equipes**: Gestão de membros da equipe e funções\n    * **Clientes**: Gestão de clientes e níveis de comunicação\n    * **Riscos**: Análise e gestão de riscos com IA\n    * **Checklists**: Sistema de checklists para projetos\n    * **Action Items**: Gestão de itens de ação e tarefas\n    \n    ### Arquitetura:\n    \n    * **Services Layer**: Lógica de negócio separada dos controllers\n    * **Repository Pattern**: Camada de abstração para acesso a dados\n    * **Cache Redis**: Sistema de cache para performance\n    * **Processamento Assíncrono**: Tarefas em background com Celery\n    * **IA Integration**: Integração com Gemini API para análise de riscos\n    \n    ### Tecnologias:\n    \n    * **FastAPI**: Framework web moderno e rápido\n    * **SQLAlchemy**: ORM robusto com suporte assíncrono\n    * **PostgreSQL**: Banco de dados relacional\n    * **Redis**: Sistema de cache\n    * **Celery**: Processamento assíncrono\n    * **Pydantic**: Validação de dados\n    ",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/action-items": {
      "get": {
        "tags": [
          "action-items"
        ],
        "summary": "List Action Items",
        "operationId": "list_action_items_api_v1_action_items_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ActionItemOut"
                  },
                  "type": "array",
                  "title": "Response List Action Items Api V1 Action Items Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "action-items"
        ],
        "summary": "Create Action Item",
        "operationId": "create_action_item_api_v1_action_items_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionItemIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionItemOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/action-items/{action_item_id}": {
      "get": {
        "tags": [
          "action-items"
        ],
        "summary": "Get Action Item",
        "operationId": "get_action_item_api_v1_action_items__action_item_id__get",
        "parameters": [
          {
            "name": "action_item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Action Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionItemOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActionItemIn": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "project_id": {
            "type": "integer",
            "title": "Project Id"
          },
          "priority": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority"
          },
          "assignee_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assignee Id"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status",
            "default": "open"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "due_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Date"
          }
        },
        "type": "object",
        "required": [
          "title",
          "type",
          "project_id"
        ],
        "title": "ActionItemIn"
      },
      "ActionItemOut": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "project_id": {
            "type": "integer",
            "title": "Project Id"
          },
          "priority": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority"
          },
          "assignee_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assignee Id"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status",
            "default": "open"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "due_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Date"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "title",
          "type",
          "project_id",
          "id",
          "created_at"
        ],
        "title": "ActionItemOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}